No code: Add e2e tests
