+++ test/unit/test_magic.py
