+++ test/unit/test_paulis_with_placeholders.py
