+++ test/unit/test_rewrite_arithmetic.py
